Add SLURM script for training and update default context format to 'long'

+++ scripts/4_training_decoder/build_oof_memory_dataset.py
@@ -520,7 +520,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--context-format",
         type=str,
-        default="short",
+        default="long",
         choices=["short", "long", "hybrid_short", "hybrid_long", "hybrid_medium"],
         help="Context format used in train pickle filenames",
     )
